backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-api-python-client>=2.100.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
